fix tests

+++ src/test/kotlin/ar/edu/unq/desapp/grupoc/backenddesappapi/model/CryptoCurrencyTests.kt
@@ -4,6 +4,7 @@ import ar.edu.unq.desapp.grupoc.backenddesappapi.model.enums.SYMBOL
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 class CryptoCurrencyTests {
@@ -11,13 +12,13 @@ class CryptoCurrencyTests {
     @Test
     fun shouldShowCryptoCurrencyInfo() {
         val crypto = CryptoCurrency(
-            price = 100000.0,
+            price = BigDecimal.valueOf(100000.0),
             symbol = SYMBOL.BTCUSDT,
             lastUpdateDateAndTime = LocalDateTime.now()
         )
 
         assertAll( {
-            Assertions.assertEquals(crypto.price, 100000.0);
+            Assertions.assertEquals(crypto.price, BigDecimal.valueOf(100000.0));
             Assertions.assertEquals(crypto.symbol, SYMBOL.BTCUSDT);
             Assertions.assertNotNull(crypto.lastUpdateDateAndTime)
         })
